#231 - Moved StorageLock to com.artipie.asto.lock.storage
